Refactor UserService::getUnviewedQuotesForUser
This change refactors the method to be able to return both viewed and
unviewed quotes for a user. After reflecting on the method, while
considering adding a second method for returning viewed quotes, it made
little sense to me to add a second method, when adding an extra
parameter and renaming the method could achieve the same affect. An enum
was used to cut down on required code to implement the change.

It also adds several further quote authors and quotes so that the tests
are a bit more meaningful.