Add mezzio-flash and mezzio-twigrenderer as dependencies
These packages are required to provide HTML template support and flash
message support to the application. I could have used the slim/twig-view
component instead, which is, perhaps, more "native" to Slim
applications. But I'm more familiar with these packages, and the choice
is mine to architect the application as makes most sense to me.