Add flash message support
This change adds the ability to set a flash message before the request
is redirected back to the subscribe form. If the process was successful,
then the user is notified with a success message. If the process was not
successful, then the user is notified with a failure message. The code
flow was streamlined to have it be as simple as possible as well.