Refactor mobile number validation in SubscribeByMobileHandler
I'd been meaning to complete this refactor for a little while now, but
put it off while working on other functionality. It refactors
SubscribeByMobileHandler to use MobileNumberInputFilter to validate the
provided mobile number, instead of using preg_match directly. The intent
of this change is to make mobile number validation universal across the
application, increasing maintainability.