Move the subscribe and unsubscribe route definitions to the email trait
I've been meaning to do this for some time, that is, to store the routes
in as few places as possible. Given the way the routing table's
constructed, it's not possible, I believe, to reduce it down completely
to only one location. So this is as close as I can do it.