Make some class properties readonly
I've been learning ever more about immutable classes and have been
coming to appreciate it as a positive addition in a number of instances.
To me, the classes that this change affects are such instances. The
properties will never change, so it makes sense to me to, at the
language level, enforce that.