Refactor successful mobile subscribe and unsubscribes
After reading through the respective Twilio documentation and chatting
about it in the respective Slack channel, TwiML doesn't need to be sent
in response to a webhook. It's optional. However, it's still good form
to send an HTTP 204 if the operation triggered by the webhook was
successful.

Given that, I'm making this change to refactor the code accordingly.