Add ability to return a printable quote
This isn't the best of names, but this change adds the ability to return
what I'm calling a "printable quote". That is, a quote including both
the quote text as well as the author's name, separated by a hyphen,
which is typically how they're rendered. That way, it doesn't need to be
formatted later. This might be a break of convention, i.e, this should
be done in a view layer. But I wanted to provide this functionality
without adding too much code overhead. Move it to a better place, if
relevant.