Refactor how the Command classes are configured
After reading further through the relevant documentation, and
encountering an error because the name wasn't set in either Command
class, I've added this change. By using PHP 8 attributes, it simplifies
setting the relevant class properties. The change also moves the
configure method to just after the class constructor, because that's
just where I want to have it; no other reason. It also adds some nominal
output, so that the user will know if something is happening or not.